WEB/backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
langchain-community
ollama
pydantic
